Revert "refactored hold_forks + attempting to fix 1 philo problem"
This reverts commit cb272ad1be5dabadf71cc1c001465a03952b2f9d.

+++ src/time.c
@@ -6,14 +6,14 @@
 /*   By: rvan-duy <rvan-duy@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/04/03 12:46:51 by rvan-duy      #+#    #+#                 */
-/*   Updated: 2022/04/03 15:02:25 by rvan-duy      ########   odam.nl         */
+/*   Updated: 2022/04/03 12:51:22 by rvan-duy      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <sys/time.h>
 #include <unistd.h>
 
-size_t	get_time_in_ms(void)
+long long	get_time_in_ms(void)
 {
 	struct timeval	time;
 
@@ -21,15 +21,15 @@ size_t	get_time_in_ms(void)
 	return (time.tv_sec * 1000LL + time.tv_usec / 1000);
 }
 
-size_t	get_timestamp(size_t start_timestamp)
+long long	get_timestamp(long long start_timestamp)
 {
 	return (get_time_in_ms() - start_timestamp);
 }
 
-static size_t	get_time_us(void)
+static unsigned long	get_time_us(void)
 {
 	static struct timeval	tv;
-	size_t					out;
+	unsigned long			out;
 
 	gettimeofday(&tv, NULL);
 	out = tv.tv_sec;
@@ -38,9 +38,9 @@ static size_t	get_time_us(void)
 	return (out);
 }
 
-void	stupid_sleep(size_t time2sleep)
+void	stupid_sleep(unsigned long time2sleep)
 {
-	size_t	start;
+	unsigned long	start;
 
 	start = get_time_us();
 	time2sleep *= 1000;
